Refactor renderers and expose CSV headers
