+++ src/lib.rs
@@ -6,7 +6,8 @@ pub mod migrate;
 pub mod models;
 pub mod parser;
 pub mod postprocess;
-pub mod renderer;
+pub mod renderers;
+pub use renderers as renderer;
 pub mod schema;
 pub mod template;
 pub mod templates;
