+++ src/main.rs
@@ -18,7 +18,8 @@ mod models;
 mod parser;
 mod plugin_index;
 mod postprocess;
-mod renderer;
+mod renderers;
+use renderers as renderer;
 mod schema;
 mod template;
 mod templates;
